requirements.txt
chromadb
openai
streamlit
streamlit-lottie
python-dotenv
sentence-transformers
numpy
scikit-learn
numpy
torch
transformers
langchain-core
langchain-community
tiktoken
bs4